Update radical_onyomi_complete_map.sql
Added logic to remove duplicate radical combinations
e.g. for the kanji 繢聵憒嘳蕢, the common radical is the thing on right which is somehow a combination of ｜,ハ,一,口,目,虫,貝; but old logic also captured the combinations: ハ,一,目,貝 and｜,ハ,目,貝 and ハ,口,目,貝 and ｜,ハ,一,目,貝. The new logic only fetches the longest one, I can trim it down manually later. This narrows list down from 1800 to 600